[#17765] Consolidate JUnit categories and tags into testing/jupiter/tags
