+++ pom.xml
@@ -1215,11 +1215,6 @@
             <version>${version.arquillian}</version>
          </dependency>
          <dependency>
-            <groupId>org.jboss.arquillian.junit5</groupId>
-            <artifactId>arquillian-junit5-container</artifactId>
-            <version>${version.arquillian}</version>
-         </dependency>
-         <dependency>
             <groupId>org.jboss.arquillian.testng</groupId>
             <artifactId>arquillian-testng-container</artifactId>
             <version>${version.arquillian}</version>
