+++ server/rollingupgradetests/src/test/java/org/infinispan/server/rollingupgrade/RollingUpgradeAuthenticationIT.java
@@ -6,9 +6,9 @@ import org.infinispan.server.security.authentication.RespAuthentication;
 import org.infinispan.server.security.authentication.RestAuthentication;
 import org.infinispan.server.test.api.TestUser;
 import org.infinispan.server.test.core.rollingupgrade.RollingUpgradeConfigurationBuilder;
-import org.infinispan.server.test.core.tags.Security;
 import org.infinispan.server.test.jupiter.InfinispanSuite;
 import org.infinispan.server.test.jupiter.RollingUpgradeHandlerExtension;
+import org.infinispan.testing.jupiter.tags.Security;
 import org.junit.jupiter.api.extension.RegisterExtension;
 import org.junit.platform.suite.api.SelectClasses;
 import org.junit.platform.suite.api.Suite;
