+++ server/tests/src/test/java/org/infinispan/server/persistence/AsyncJdbcStringBasedCacheStore.java
@@ -12,7 +12,7 @@ import org.infinispan.server.test.jupiter.InfinispanServer;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ArgumentsSource;
 
-@org.infinispan.server.test.core.tags.Database
+@org.infinispan.testing.jupiter.tags.Database
 public class AsyncJdbcStringBasedCacheStore {
 
     @InfinispanServer(PersistenceIT.class)
