+++ server/tests/src/test/java/org/infinispan/server/persistence/BaseJdbcStringBasedCacheStore.java
@@ -12,7 +12,7 @@ import org.infinispan.testing.Eventually;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ArgumentsSource;
 
-@org.infinispan.server.test.core.tags.Database
+@org.infinispan.testing.jupiter.tags.Database
 public class BaseJdbcStringBasedCacheStore {
 
    @InfinispanServer(PersistenceIT.class)
