+++ server/tests/src/test/java/org/infinispan/server/persistence/JdbcPing2IT.java
@@ -17,10 +17,10 @@ import org.infinispan.commons.dataconversion.internal.Json;
 import org.infinispan.server.test.core.ServerRunMode;
 import org.infinispan.server.test.core.persistence.ContainerDatabase;
 import org.infinispan.server.test.core.persistence.DatabaseServerListener;
-import org.infinispan.server.test.core.tags.Database;
 import org.infinispan.server.test.jupiter.InfinispanServerExtension;
 import org.infinispan.server.test.jupiter.InfinispanServerExtensionBuilder;
 import org.infinispan.testing.Eventually;
+import org.infinispan.testing.jupiter.tags.Database;
 import org.junit.jupiter.api.condition.DisabledIf;
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.api.extension.RegisterExtension;
