+++ server/tests/src/test/java/org/infinispan/server/persistence/RocksDBStoreIT.java
@@ -7,9 +7,9 @@ import org.infinispan.client.hotrod.RemoteCache;
 import org.infinispan.commons.configuration.StringConfiguration;
 import org.infinispan.commons.util.Version;
 import org.infinispan.server.test.core.ServerRunMode;
-import org.infinispan.server.test.core.tags.Persistence;
 import org.infinispan.server.test.jupiter.InfinispanServerExtension;
 import org.infinispan.server.test.jupiter.InfinispanServerExtensionBuilder;
+import org.infinispan.testing.jupiter.tags.Persistence;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
