+++ server/tests/src/test/java/org/infinispan/server/resilience/ResilienceMaxRetryIT.java
@@ -18,10 +18,10 @@ import org.infinispan.client.hotrod.configuration.ConfigurationBuilder;
 import org.infinispan.client.hotrod.exceptions.TransportException;
 import org.infinispan.client.hotrod.impl.transport.netty.OperationDispatcher;
 import org.infinispan.configuration.cache.CacheMode;
-import org.infinispan.server.test.core.tags.Resilience;
 import org.infinispan.server.test.jupiter.InfinispanServerExtension;
 import org.infinispan.server.test.jupiter.InfinispanServerExtensionBuilder;
 import org.infinispan.test.TestingUtil;
+import org.infinispan.testing.jupiter.tags.Resilience;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
