+++ server/tests/src/test/java/org/infinispan/server/security/AuthenticationLDAPSIT.java
@@ -9,9 +9,9 @@ import org.infinispan.configuration.cache.CacheMode;
 import org.infinispan.server.test.api.TestUser;
 import org.infinispan.server.test.core.Common;
 import org.infinispan.server.test.core.LdapServerListener;
-import org.infinispan.server.test.core.tags.Security;
 import org.infinispan.server.test.jupiter.InfinispanServerExtension;
 import org.infinispan.server.test.jupiter.InfinispanServerExtensionBuilder;
+import org.infinispan.testing.jupiter.tags.Security;
 import org.junit.jupiter.api.extension.RegisterExtension;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ArgumentsSource;
