+++ server/tests/src/test/java/org/infinispan/server/security/authentication/RestAuthentication.java
@@ -23,9 +23,9 @@ import org.infinispan.client.rest.configuration.Protocol;
 import org.infinispan.client.rest.configuration.RestClientConfigurationBuilder;
 import org.infinispan.server.test.api.TestClientDriver;
 import org.infinispan.server.test.core.Common;
-import org.infinispan.server.test.core.tags.Security;
 import org.infinispan.server.test.jupiter.InfinispanServer;
 import org.infinispan.testing.Exceptions;
+import org.infinispan.testing.jupiter.tags.Security;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.params.ParameterizedTest;
