+++ server/tests/src/test/java/org/infinispan/server/security/authorization/AuthorizationKerberosIT.java
@@ -13,10 +13,10 @@ import org.infinispan.server.test.artifacts.Artifacts;
 import org.infinispan.server.test.core.Common;
 import org.infinispan.server.test.core.KerberosListener;
 import org.infinispan.server.test.core.LdapServerListener;
-import org.infinispan.server.test.core.tags.Security;
 import org.infinispan.server.test.jupiter.InfinispanServerExtension;
 import org.infinispan.server.test.jupiter.InfinispanServerExtensionBuilder;
 import org.infinispan.server.test.jupiter.InfinispanSuite;
+import org.infinispan.testing.jupiter.tags.Security;
 import org.junit.jupiter.api.extension.RegisterExtension;
 import org.junit.platform.suite.api.SelectClasses;
 import org.junit.platform.suite.api.Suite;
