--- server/tests/src/test/java/org/infinispan/server/test/core/category/Persistence.java
@@ -1,8 +1,0 @@
-package org.infinispan.server.test.core.category;
-
-/**
- * @author Tristan Tarrant &lt;tristan@infinispan.org&gt;
- * @since 10.0
- **/
-public class Persistence {
-}
