--- testing/src/main/java/org/infinispan/testing/junit/categories/Profiling.java
@@ -1,7 +1,0 @@
-package org.infinispan.testing.junit.categories;
-
-/**
- * JUnit category for profiling tests.
- */
-public interface Profiling {
-}
