--- testing/src/main/java/org/infinispan/testing/junit/categories/Smoke.java
@@ -1,7 +1,0 @@
-package org.infinispan.testing.junit.categories;
-
-/**
- * JUnit category for smoke tests.
- */
-public interface Smoke {
-}
