--- testing/src/main/java/org/infinispan/testing/junit/categories/Stress.java
@@ -1,7 +1,0 @@
-package org.infinispan.testing.junit.categories;
-
-/**
- * JUnit category for stress tests.
- */
-public interface Stress {
-}
