--- testing/src/main/java/org/infinispan/testing/junit/categories/Unstable.java
@@ -1,7 +1,0 @@
-package org.infinispan.testing.junit.categories;
-
-/**
- * JUnit category for unstable tests.
- */
-public interface Unstable {
-}
