+++ testing/src/main/java/org/infinispan/testing/jupiter/tags/Database.java
@@ -1,0 +1,16 @@
+package org.infinispan.testing.jupiter.tags;
+
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+import org.junit.jupiter.api.Tag;
+
+@Target({ ElementType.TYPE, ElementType.METHOD })
+@Retention(RetentionPolicy.RUNTIME)
+@Tag("database")
+@Tag("persistence")
+public @interface Database {
+}
