+++ testing/src/main/java/org/infinispan/testing/jupiter/tags/Resilience.java
@@ -1,0 +1,15 @@
+package org.infinispan.testing.jupiter.tags;
+
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+import org.junit.jupiter.api.Tag;
+
+@Target({ ElementType.TYPE, ElementType.METHOD })
+@Retention(RetentionPolicy.RUNTIME)
+@Tag("resilience")
+public @interface Resilience {
+}
