+++ testing/src/main/java/org/infinispan/testing/jupiter/tags/Smoke.java
@@ -1,5 +1,4 @@
-package org.infinispan.server.test.core.tags;
-
+package org.infinispan.testing.jupiter.tags;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -10,6 +9,6 @@ import org.junit.jupiter.api.Tag;
 
 @Target({ ElementType.TYPE, ElementType.METHOD })
 @Retention(RetentionPolicy.RUNTIME)
-@Tag("resilience")
-public @interface Resilience {
+@Tag("smoke")
+public @interface Smoke {
 }
